Main

+++ Bot/bot.js
@@ -26,9 +26,21 @@ bot.on('message', msg => {
 			case "cleanMsgs":
 				if(authenticate("Admin",msg)){
 					msg.channel.fetchMessages().then(function(data){
-							if(args[0]==="all"){
-								msg.channel.bulkDelete(data);	
-							}else{
+							if(args.length>0){
+								if(args[0]==="all"){
+									msg.channel.bulkDelete(data);	
+								}
+								else{
+									var msgs = data.array();
+									var num=parseInt(args[0]);
+									for(var i=0;i<msgs.length && i<=num;i++){
+										if(!msgs[i].pinned){
+											msgs[i].delete();
+										}
+									}
+								}
+							}
+							else{
 								var msgs = data.array();
 								for(var i=0;i<msgs.length;i++){
 									if(!msgs[i].pinned){
@@ -38,15 +50,11 @@ bot.on('message', msg => {
 							}
 							
 					});
-				}else{
-					msg.reply("You must have the \"Admin\" role to use this command");
 				}
 				break;		
 			case "botinfo":
 				if(authenticate("Admin",msg)){
 					msg.reply("\nClient: "+bot.user+"\nPing: "+bot.ping+"\nStatus: "+bot.status+"\nUptime: "+(bot.uptime/1000) +" (secs)"+"\nReady at: "+bot.readyAt);
-				}else{
-					msg.reply("You must have the \"Admin\" role to use this command");
 				}
 				break;
 			default: msg.reply("Unrecognized Command");   
@@ -66,11 +74,11 @@ function authenticate(roleName,msg){
 		return;
 	}
 	let role = msg.guild.roles.find("name", roleName);
-	if(role===null){
-		return false;
+	if(role!=null){
+		if(msg.member.roles.has(role.id)) {
+			return true;
+		}
 	}
-	if(msg.member.roles.has(role.id)) {
-		return true;
-	}
+	msg.reply("You must have the \""+roleName+"\" role to use this command");
 	return false;
 }
